requirements.txt
fastapi==0.86.0
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.6.0
